--- .eslintignore
@@ -1,3 +1,0 @@
-*.css
-*.svg
-*.scss.d.ts
