+++ package.json
@@ -69,7 +69,8 @@
         "react-helmet": "^6.1.0",
         "react-helmet-async": "^1.3.0",
         "react-loader-spinner": "^5.4.5",
-        "react-router-dom": "^6.7.0",
+        "react-router": "^6.15.0",
+        "react-router-dom": "^6.15.0",
         "terser": "^5.15.1",
         "unplugin-swc": "^1.3.2",
         "vite": "^3.2.3",
