+++ src/pages/app/App.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Routes, Route, HashRouter } from 'react-router-dom';
+import { initializeI18n } from '../../common/locales/initialize-i18n';
+
 import { NavBar } from '../../components/navbar/NavBar';
-import { NavMenuItem } from '../../common/interfaces/SharedInterfaces';
 import { MainPage } from '../main-page/MainPage';
 import {
     SocialMediaComponent
@@ -12,62 +13,37 @@ import { SponsorsPage } from '../sponsors-page/SponsorsPage';
 import { AboutUs } from '../about-us/AboutUs';
 import './app.module.scss';
 import { JoinUs } from '../join-us/JoinUs';
-import Cookies from 'js-cookie';
-import i18n from 'i18next';
-import LanguageDetector from 'i18next-browser-languagedetector';
-import en from '../../common/locales/en/defaults.json';
-import pl from '../../common/locales/pl/defaults.json';
-import localConfig from '../../../vite.local.config';
-
-interface NavMenuModel extends NavMenuItem {
-    component: () => React.ReactElement;
-}
-
-export const initi18n = () => {
-    const langCookieKey = 'i18next';
-    const storedLang = Cookies.get(langCookieKey) ?? 'pl';
-
-    i18n.use(LanguageDetector).init({
-        fallbackLng: 'pl',
-        resources: {
-            en: { translation: en },
-            pl: { translation: pl }
-        },
-        lng: storedLang,
-        detection: {
-            lookupCookie: langCookieKey,
-            caches: ['cookie'],
-            cookieMinutes: 60*24*30
-        }
-    });
-
-    window.__ = (str: string) => i18n.t(str);
-    window.classes = (...args: string[]) => args.join(' ');
-    if (!window.appContext) {
-        window.appContext = {
-            baseUrl: (localConfig as any).base ?? '/'
-        };
-    }
-};
+import { NavMenuItem } from '../../common/interfaces/SharedInterfaces';
 
 const App = () => {
+    initializeI18n();
 
-    initi18n();
-
-    const getNavMenuModel = (name: string, url: string, component: React.ReactElement) : NavMenuModel =>  {
-        return {
-            name: name,
-            url: url,
-            component: () => component
-        };
-    };
-
-    const menuItems: NavMenuModel[] = [
-        getNavMenuModel(__('navbar.home'), '/', <MainPage/>),
-        getNavMenuModel(__('navbar.aboutUs'), '/about-us', <AboutUs/>),
-        getNavMenuModel(__('navbar.joinUs'), '/join-us', <JoinUs/>),
-        getNavMenuModel(__('navbar.sponsors'), '/sponsors', <SponsorsPage/>),
-        getNavMenuModel(__('navbar.contact'), '/contact', <ContactPage/>),
+    const menuItems: NavMenuItem[] = [
+        {
+            name: __('navbar.home'),
+            url: '/',
+            component: () => <MainPage/>
+        },
+        {
+            name: __('navbar.aboutUs'),
+            url: '/about-us',
+            component: () => <AboutUs/>
+        },
+        {
+            name: __('navbar.joinUs'),
+            url: '/join-us',
+            component: () => <JoinUs/>
+        },
+        {
+            name: __('navbar.sponsors'),
+            url: '/sponsors',
+            component: () => <SponsorsPage/>
+        },
+        {
+            name: __('navbar.contact'),
+            url: '/contact',
+            component: () => <ContactPage/>
+        }
     ];
 
     return (
